updating create team validation

+++ controllers/serviceController.js
@@ -1,3 +1,4 @@
+const Team = require('../models/teamModel');
 const AppError = require('../utils/appError');
 const catchAsync = require('../utils/catchAsync');
 const Service = require('./../models/serviceModel');
@@ -19,7 +20,10 @@ exports.createService = catchAsync(async (req, res, next) => {
     return next(new AppError('Service durations are required!', 400));
   }
 
-  const newService = await Service.create({ ...req.body, user: req.user._id });
+  const newService = await Service.create({
+    ...req.body,
+    creator: req.user._id,
+  });
 
   res.status(201).json({
     status: 'success',
@@ -30,10 +34,13 @@ exports.createService = catchAsync(async (req, res, next) => {
 });
 
 exports.getService = catchAsync(async (req, res, next) => {
-  const service = await Service.findById(req.params.id);
+  const service = await Service.findById(req.params.id).populate(
+    'creator',
+    'firstName lastName photo'
+  );
 
   if (!service) {
-    return next(new AppError('No service found with that ID!'));
+    return next(new AppError('No service found with that ID!', 404));
   }
 
   res.status(200).json({
@@ -47,7 +54,7 @@ exports.getService = catchAsync(async (req, res, next) => {
 exports.updateService = catchAsync(async (req, res, next) => {
   const service = await Service.findById(req.params.id);
   if (!service) {
-    return next(new AppError('No service found with that ID!'));
+    return next(new AppError('No service found with that ID!', 404));
   }
 
   if (req.body.durations && req.body.durations.length === 0) {
@@ -58,12 +65,13 @@ exports.updateService = catchAsync(async (req, res, next) => {
     req.params.id,
     req.body,
     {
+      new: true,
       runValidators: true,
     }
   );
 
   if (!updatedService) {
-    return next(new AppError('No service found with that ID!'));
+    return next(new AppError('No service found with that ID!', 404));
   }
 
   res.status(200).json({
@@ -75,11 +83,20 @@ exports.updateService = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteService = catchAsync(async (req, res, next) => {
-  const service = await Service.findByIdAndDelete(req.params.id);
+  const service = await Service.findById(req.params.id);
 
   if (!service) {
-    return next(new AppError('No service found with that ID!'));
+    return next(new AppError('No service found with that ID!', 404));
   }
+
+  const teams = await Team.find({ serviceHours: service._id });
+  if (teams.length > 0) {
+    return next(
+      new AppError("Couldn't delete services already used in teams!", 400)
+    );
+  }
+
+  await Service.findByIdAndDelete(req.params.id);
 
   res.status(200).json({
     status: 'success',
